refactor(site): extract shared provider-state logic into aiModels module
Move the provider-state derivation and canManageProviderModels helper
into modules/aiModels/providerStates (with tests) and replace the
duplicated copies in the legacy ChatModelAdminPanel. No behavior change.
